feat: transfer writeLogToFile func into separate from logger file
